@wasz we need to change the keywords: “ Margesic” to “Margesic A-C” “Codimal” to “codimal ph”, “codimal bn”, codimal dh”, --à comidal, codimal a, codimal dm, and codimal la are not opioids “Stagesic” to “Stagesic-10” “lortuss” to “lortuss hc” “lortuss ex” “relcof” to “relcof c” yet it still would then pull “relcof cpm” which should not be included “Histinex” to “histinex hc” “histinex d” , but “histinex dm” should not be included.
Also for a lot of the brand names like “poly-tussin”, “de-chlor”, and “tussar” them by themselves are opioids but if they have DM after it is not an opioid. Not sure how we can build this language in. The name brands are seeming like the biggest problem for us.

“Ultram”  is a name brand for tramadol, yet it is bring up a lot of non-opioid so we may need to add those separately
I am very confused about Dolorex because online it says it is an opioid brand name, but it is pull up the brand name for some non-opioids I think à I think we need to check this one after manually or get an MD to confer with us and whether we should include it as a keyword
I think we need to get either get rid of “Fioricet” as a key word or change it as it by itself is not bringing up opioids. Same thing with “stagesic”.